requirements.txt
docopt==0.6.2
flake8==3.8.4
mccabe==0.6.1
Pillow==8.1.2
pycodestyle==2.6.0
pyflakes==2.2.0
